+++ R/install.R
@@ -196,9 +196,10 @@ install_hugo_bin = function(exec, version) {
 #' @param version A version number. The default is to automatically detect the
 #'   latest version. Versions before v0.17 are not supported.
 #' @return A data frame containing columns \code{os} (operating system),
-#'   \code{arch} (architecture), and \code{extended} (extended version or not).
-#'   If your R version is lower than 4.1.0, a character vector of the installer
-#'   filenames will be returned instead.
+#'   \code{arch} (architecture), \code{extended} (extended version or not),
+#'   \code{withdeploy} (with the deploy feature or not). If your R version is
+#'   lower than 4.1.0, a character vector of the installer filenames will be
+#'   returned instead.
 #' @export
 #' @examplesIf interactive()
 #' blogdown::hugo_installers()
@@ -206,7 +207,7 @@ install_hugo_bin = function(exec, version) {
 #' blogdown::hugo_installers('0.17')
 hugo_installers = function(version = 'latest') {
   repo = 'gohugoio/hugo'
-  if (version == 'latest') version = xfun::github_releases(repo, 'latest')
+  if (version == 'latest') version = xfun::github_releases(repo, 'latest')[1]
   version = sub('^[vV]?', 'v', version)
   json = xfun::loadable('jsonlite')
   res = xfun::github_api(sprintf('/repos/%s/releases/tags/%s', repo, version), raw = !json)
@@ -216,19 +217,20 @@ hugo_installers = function(version = 'latest') {
     res = strsplit(res, '"browser_download_url":"')
     xfun::grep_sub('^(https://[^"]+)".*', '\\1', unlist(res))
   }
-  res = grep('[.](zip|tar[.]gz)$', unlist(res), value = TRUE)
+  res = grep('[.](zip|tar[.]gz|pkg)$', unlist(res), value = TRUE)
   res = basename(res)
   if (!'gregexec' %in% ls(baseenv())) {
     warning('Your R version is too low (< 4.1.0) and does not have gregexec().')
     return(res)
     gregexec = regexec  # a hack to pass R CMD check without NOTE
   }
-  m = gregexec('^hugo_(extended_)?([^_]+)_([^-]+)-([^.]+)[.](zip|tar[.]gz)$', res)
-  res = lapply(regmatches(res, m), function(x) if (length(x) >= 6) x[c(1, 3, 4, 5, 2)])
+  m = gregexec('^hugo_(extended_)?(withdeploy_)?([^_]+)_([^-]+)-([^.]+)[.]([a-z.]+)$', res)
+  res = lapply(regmatches(res, m), function(x) if (length(x) >= 6) x[c(1, 4, 5, 6, 2, 3)])
   res = do.call(rbind, res)
-  colnames(res) = c('installer', 'version', 'os', 'arch', 'extended')
+  colnames(res) = c('installer', 'version', 'os', 'arch', 'extended', 'withdeploy')
   res = as.data.frame(res)
   res$extended = res$extended == 'extended_'
+  res$withdeploy = res$withdeploy == 'withdeploy_'
   rownames(res) = res$installer
   res = res[, -1]
   res